par.sed: handle negative addresses
Might want to simplify the generated code at some point, this generates
a xor true to invert the left operand truth value, which isn't as
readable as `! value`.
The issue lies in the opposite parsing order for the C and sed:

C   -> if (!value) { [...]
sed -> /value/!{ [...]

which makes generating the negation at the beginning harder, hence
using: if (value ^ true) { [...]