misc: compile regexes only once
This is a big change. The previous design was currently compiling a
regex each time the script was going through it, meaning if my script
was s/foo/bar/ and my input was a million lines, then we would compile
foo a million times. Now we're compiling once, which has improved the
performance tremendously.

The generated code is now much harder to understand though, but there is
room for improvement.