compat: skip lookahead read when no last addr
The only way to handle $ (last line addr) is to preemptively read the
next line, and if we fail we know that the current line is the last one.
Issue with that is interactive mode, we would basically be always one
read late, meaning an action given on the readline will always happen
after the next read, which is basically unusable.

The trick GNU and bsd sed use is to optimize and skip this lookahead
read when no $ address is used. I wasn't sure if this was something I
wanted to do, considering it makes no sense from the user point of view,
it is a hack to circumvent an implementation detail, but in the end I
really wanted to be able to use my tic tac toe (added in samples).

Example below with GNU sed:

  sh$ sed 's/^/-> You wrote: /'
  Foo
  -> You wrote: Foo
  Bar
  -> You wrote: Bar
  Funk
  -> You wrote: Funk

  ----

  sh$ sed 's/^/-> You wrote: /; $s/$/ Bye!/'
  Foo
  Bar
  -> You wrote: Foo
  Funk
  -> You wrote: Bar       <-----press ctrl-d after this one
  -> You wrote: Funk Bye!